Agregado boton busqueda mas funcionalidad
